Fix MIC rebuild detection, LJ allocation, and test warnings
- Add PBC-aware rebuild detection kernels and bindings (Warp, Torch, JAX)
- Default max_nbins is 8192
- Pass cell_inv to rebuild check in MD benchmark/example infra to use
  minimum-image convention, fixing spurious rebuilds at periodic boundaries
- Sync wp_cell_inv in NPT/NPH force callbacks after cell updates
- Add pre-allocated output buffers (energies_out, forces_out, virial_out)
  to lj_energy_forces to avoid per-step allocation
- Fix JAX int64-to-int32 scatter FutureWarning across 7 neighbor files

+++ nvalchemiops/jax/neighbors/batch_naive.py
@@ -270,7 +270,7 @@ def batch_naive_neighbor_list(
         max_neighbors = estimate_max_neighbors(cutoff)
 
     if fill_value is None:
-        fill_value = positions.shape[0]
+        fill_value = jnp.int32(positions.shape[0])
 
     if neighbor_matrix is None:
         neighbor_matrix = jnp.full(
@@ -284,7 +284,7 @@ def batch_naive_neighbor_list(
     if num_neighbors is None:
         num_neighbors = jnp.zeros(positions.shape[0], dtype=jnp.int32)
     elif rebuild_flags is None:
-        num_neighbors = num_neighbors.at[:].set(0)
+        num_neighbors = num_neighbors.at[:].set(jnp.int32(0))
 
     if pbc is not None:
         if neighbor_matrix_shifts is None:
@@ -293,7 +293,7 @@ def batch_naive_neighbor_list(
                 dtype=jnp.int32,
             )
         elif rebuild_flags is None:
-            neighbor_matrix_shifts = neighbor_matrix_shifts.at[:].set(0)
+            neighbor_matrix_shifts = neighbor_matrix_shifts.at[:].set(jnp.int32(0))
         if (
             total_shifts is None
             or shift_offset is None
