+++ nvalchemiops/jax/neighbors/neighbor_utils.py
@@ -139,7 +139,7 @@ def compute_naive_num_shifts(
 
     # Compute cumulative sum for shift offsets (pure JAX, jit-compatible)
     shift_offset = jnp.zeros(num_systems + 1, dtype=jnp.int32)
-    shift_offset = shift_offset.at[1:].set(jnp.cumsum(num_shifts))
+    shift_offset = shift_offset.at[1:].set(jnp.cumsum(num_shifts, dtype=jnp.int32))
 
     # NOTE: total_shifts_value requires int() extraction - this is needed for
     # array allocation and must be concrete. The caller needs to handle this.
@@ -230,7 +230,7 @@ def get_neighbor_list_from_neighbor_matrix(
 
     # Create CSR-style pointer array
     neighbor_ptr = jnp.zeros(num_neighbors.shape[0] + 1, dtype=jnp.int32)
-    neighbor_ptr = neighbor_ptr.at[1:].set(jnp.cumsum(num_neighbors))
+    neighbor_ptr = neighbor_ptr.at[1:].set(jnp.cumsum(num_neighbors, dtype=jnp.int32))
 
     if neighbor_shift_matrix is not None:
         neighbor_list_shifts = neighbor_shift_matrix[mask]
@@ -308,7 +308,9 @@ def prepare_batch_idx_ptr(
             batch_idx, minlength=num_systems, length=num_systems
         )
         batch_ptr = jnp.zeros(num_systems + 1, dtype=jnp.int32)
-        batch_ptr = batch_ptr.at[1:].set(jnp.cumsum(num_atoms_per_system))
+        batch_ptr = batch_ptr.at[1:].set(
+            jnp.cumsum(num_atoms_per_system, dtype=jnp.int32)
+        )
 
     return batch_idx, batch_ptr
